requirements.txt
Django==1.8
pymongo==2.8
mongoengine==0.9.0
Pillow==2.7.0
